Set appropriate flags as required or mutually exclusive  (#26)
# Description of change
Explicitly marking required flags as required has some user benefits. For example:
1. The program fast-fails with a helpful message if a required flag is not set e.g.-
`required flag(s) "description", "location", "name" not set`
2. Autocomplete will generate a list of required flags on `tab` `tab` e.g.-
```
--description (The description of the experience)  --name (The name of the experience)
--location (The location of the experience, e.g. an S3 URI for the experience folder)
```

Unfortunately Cobra does not currently indicate that flags are required in the usage string
produced when `--help` is typed. There is a ticket for this with an open PR:
https://github.com/spf13/cobra/issues/1537

Two possible work arounds we could consider in a follow-up PR are:
1. Manually injecting the terms `[required]` or `[optional]` into the usage strings.
2. Using `SetUsageTemplate` to apply a custom template. Although, this is not as trivial
as it sounds. Nonetheless we could crib the PR in the ticket linked above.

## Guide to reproduce test results
Build and enable autocomplete - as described in the docs.
Try `tab` `tab` on commands to see the required flags.
Try `return` on commands to see the required flags.

## Checklist

- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in this PR.