Getting Styley (#44)
This PR sets us up to be able customize the style of help, usage and error 
messages. The effect of the PR is *almost* a noop on current behavior. It
makes a very small change to the styling of section headings in usage
messages (see images below), in order to demonstrate the functionality. 

The main aim of the PR is to set us up with the ability to make greater 
customization in the future. For example, the next PR (#45 ) will aim to 
separate Flags into required and optional sections per a customer request.

Old:
![image](https://github.com/resim-ai/api-client/assets/1348072/4714d634-1729-40d6-9e84-937f7f17a0bf)
New:
![image](https://github.com/resim-ai/api-client/assets/1348072/a550212a-fcac-47b8-8cc6-4d0f7941e9b0)

[Asana task](https://app.asana.com/0/1205228215063249/1205227572053881/f)
## Guide to reproduce test results
run any `resim` command with `--help` and verify the new heading styling.

## Checklist

- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in this PR.