Allow user to pass name or UUID to a generic --project flag (#46)
# Description of change
Sometime users are unsure of whether they need to pass a project ID or a project name.
This has been exacerbated by docs and our UI design. But nonetheless, the confusion
can arise. A reasonable approach might be to be permissive. That is, give them a `--project`
and allow them to pass either a name or an ID. This approach requires
a bit on consideration. But as it  is such a small change I simply created this PR, to consider. 
## Guide to reproduce test results
Try running
`resim project get --identifer <a valid project uuid or name>`
`resim project delete --identifer <a valid project uuid or name>`
`resim build create --project <a valid project uuid or name> ...`
`resim branch create --project <a valid project uuid or name> ...`

And verify that it works.
## Checklist

- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.