Properly Report Error Response (#72)
# Description of change

There was a bug in the previous versions of the CLI: any response errors were lost because the steam had already been read by the middleware. This meant when we tried to read again there was always an error "body closed". This change captures the error message and passes it through to validation steps.

## Guide to reproduce test results

N/A

## Checklist

- [X] I have self-reviewed this change.
- [X] I have tested this change.
- [X] This change is covered by tests that are already landed, or in this PR.
- [x] I have updated the changelog, if appropriate.