Use viper config file; allow persistence of project setting (#78)
# Description of change

This hooks up the config file writing-and-reading side of Viper, which
was already being used to manage flags.
When writing to the config, we first read it into a blank Viper
instance, so that we can persist different sets of flags from different
call sites into the same config, while not persisting every single flag
every time.

Also: Some minor rearrangement of the config-initialization functions in
the aid of testability.
Also also: Moved the `RegisterViperFlagsAndSetClient` init call to the
root command, where it flows down by default into every sub command.

## Guide to reproduce test results

```
$ rm ~/.resim/resim.json
$ cli project list
[... list of projects ]
$ cli branches list
2024/02/28 21:57:19 required flag(s) "project" not set
$ cli project select --project resim-sandbox-project
$ cli branches list
[
  {
    "branchID": [..list of branches continues ...]
$ cli project list
[..list of projects, again ]
* resim-sandbox-project
```

## Checklist

- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in
this PR.
- [x] I have updated the changelog, if appropriate.