Log Ingest (#109)
# Description of change

Introduces a CLI command designed to support customers who want to
ingest logs from IRL (or HiL) robot testing and run a ReSim metrics
build on it to generate results.

This is achieved by running a test batch with a single test containing
that log.

## To Do

 - [x] Support tagging the log as desired
 - [x] Support optional branch and find the main branch
 - [x] Support optional version and use a "latest" commit
 - [ ] End to end test

## Notes:

This is a proof of concept, reusing the existing infrastructure that we
have, which may or may not be the right thing to do.

## Guide to reproduce test results

```
```

## Checklist

- [ ] I have self-reviewed this change.
- [ ] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in
this PR.
- [ ] I have updated the changelog, if appropriate.