Agent <> Log Ingest (#131)
# Description of change

It is often desirable to pseudo-ingest logs and compute metrics on them
before going through the expensive process of fully ingesting into cloud
storage (particularly if there are bandwidth or cost concerns).

In this PR, we extend the log ingestion functionality of ReSim to
support pool labels so that log ingestion can occur via the ReSim Agent.

## Guide to reproduce test results

## Checklist

- [X] I have self-reviewed this change.
- [X] I have tested this change.
- [X] This change is covered by tests that are already landed, or in
this PR.
- [X] I have updated the changelog, if appropriate.