Support debug mode for Multi Container builds (#156)
# Description of change

### Feature
This adds debug mode to Multi Container builds.
When running `resim debug` you can add `--container` to put one
container into debug mode.
Other than the container in debug mode the job lifecycle is unaffected -
so if your container is not set as a [Test
Orchestrator](https://docs.resim.ai/guides/multi-container-builds/#container-lifecycle)
then it is possible that your job will come to a conclusion while you
are connected in debug mode.

### Other
This PR adds mocks for the ReSim API for use in unit testing.

## Guide to reproduce test results

<!--
    Please help reviewers by including instructions
    on how to test this change
-->

## Checklist

- [ ] I have self-reviewed this change.
- [ ] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in
this PR.
- [ ] I have updated the changelog, if appropriate.